requirements.txt
langgraph==0.3.4
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.51
langchain-openai==0.3.12
fastapi 
gunicorn
python-dotenv
chainlit==2.2.0